+++ bin/ann_flutter_flavor.dart
@@ -2,6 +2,7 @@ import 'package:args/command_runner.dart';
 import '../lib_cli/commands/sync_command.dart';
 import '../lib_cli/commands/validate_command.dart';
 import '../lib_cli/commands/firebase_command.dart';
+import '../lib_cli/commands/summary_command.dart';
 
 Future<void> main(List<String> args) async {
   final runner = CommandRunner<void>(
@@ -10,7 +11,8 @@ Future<void> main(List<String> args) async {
   )
     ..addCommand(SyncCommand())
     ..addCommand(ValidateCommand())
-    ..addCommand(FirebaseCommand());
+    ..addCommand(FirebaseCommand())
+    ..addCommand(SummaryCommand());
 
   try {
     await runner.run(args);